Use rails version 5.0
We deleted the config/* files since those were probably createt by the
rails app:update task and are not really need in an engine. Those files
should be taken care of by the application mounting the engine.

Other than that we only needed to change the version haml-rails to work
with rails 5.0 instead of rails 5.2.

Installing this commit locally with the `path:` option seems to work
with just fine in chemotion.